+++ CraueFormFlowBundle.php
@@ -2,6 +2,7 @@
 
 namespace Craue\FormFlowBundle;
 
+use Craue\FormFlowBundle\Util\TempFileUtil;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -10,4 +11,18 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class CraueFormFlowBundle extends Bundle {
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function boot() {
+		/*
+		 * Removes all temporary files created while handling file uploads.
+		 * Use a shutdown function to clean up even in case of a fatal error.
+		 */
+		register_shutdown_function(function() {
+			TempFileUtil::removeTempFiles();
+		});
+	}
+
 }